nick_model/requirements.txt
albumentations==0.4.5
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
Shapely==1.7.0
tensorflow==2.1.0
tqdm==4.46.0